app/requirements.txt
asttokens==2.0.8
attrs==22.1.0
Automat==20.2.0
backcall==0.2.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
constantly==15.1.0
cryptography==37.0.4
cssselect==1.1.0
decorator==5.1.1
et-xmlfile==1.1.0
executing==1.0.0
filelock==3.8.0
htmllaundry==2.2
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
ipdb==0.13.9
ipython==8.4.0
itemadapter==0.7.0
itemloaders==1.0.6
jedi==0.18.1
jmespath==1.0.1
lxml==4.9.1
matplotlib-inline==0.1.6
openpyxl==3.0.10
parsel==1.6.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.30
Protego==0.2.1
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyDispatcher==2.0.5
Pygments==2.13.0
pyOpenSSL==22.0.0
python-dotenv==0.20.0
queuelib==1.6.2
requests==2.28.1
requests-file==1.5.1
Scrapy==2.6.2
scrapy-xlsx==0.1.1
service-identity==21.1.0
six==1.16.0
stack-data==0.5.0
tldextract==3.3.1
toml==0.10.2
traitlets==5.3.0
Twisted==22.4.0
typing-extensions==4.3.0
urllib3==1.26.12
w3lib==2.0.1
wcwidth==0.2.5
zope.interface==5.4.0
